- Delete old admin subcommands - Move admin subcommands to main `admin` package (now that they're not several class files) - Implement the commands
