+++ bot/src/main/java/com/dongtronic/diabot/platforms/discord/commands/admin/AdminCommands.kt
@@ -1,0 +1,11 @@
+package com.dongtronic.diabot.platforms.discord.commands.admin
+
+import com.dongtronic.diabot.commands.annotations.ChildCommands
+
+@ChildCommands(
+        AdminAnnounceCommand::class,
+        AdminChannelCommands::class,
+        AdminRewardCommands::class,
+        AdminUsernameCommands::class
+)
+class AdminCommands
